fix: stabilize authentication, dashboard analytics, and UI polish
- fix: resolved infinite 401 redirect loop in Vite proxy and axios interceptor
- fix: updated PostgreSQL BigInt mapping to Java long for Ingestion and Sync counters to resolve 500 errors
- fix: corrected Bcrypt password hashes in database seed script
- feat: added UI tooltip component (TruncatedId) to cleanly display UUIDs in data tables
- feat: implemented real-time polling for the Notifications page
- chore: added backend Dockerfile and fixed docker-compose configuration
- refactor: stabilized async synchronization executor and improved logout flow

+++ backend/src/main/java/com/company/integrationplatform/ingestion/IngestionDto.java
@@ -42,13 +42,13 @@ public class IngestionDto {
         private String fileName;
 
         @Schema(description = "Total data rows in the CSV (header excluded)", example = "1500")
-        private int totalRecords;
+        private long totalRecords;
 
         @Schema(description = "Number of rows successfully processed", example = "1498")
-        private int recordsProcessed;
+        private long recordsProcessed;
 
         @Schema(description = "Number of rows that failed processing", example = "2")
-        private int recordsFailed;
+        private long recordsFailed;
 
         @Schema(description = "Job start timestamp", example = "2026-05-23T10:15:30")
         private LocalDateTime startedAt;
